printmanager/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.4
flower==2.0.1
humanize==4.12.3
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pip==24.3.1
prometheus_client==0.22.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tornado==6.5.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
